Bug 1112533 - Get rid of SVGDocumentWrapper::GetWidthOrHeight(). r=dholbert
UltraBlame original commit: 6390ede276aeb4f498d9f76ee377b5d1ef5989c0

+++ dom/svg/SVGSVGElement.h
@@ -152,8 +152,9 @@ public:
 
 
 
-  float GetIntrinsicWidth();
-  float GetIntrinsicHeight();
+
+  int32_t GetIntrinsicWidth();
+  int32_t GetIntrinsicHeight();
